Lesson 3, Task 1. change int to double for num1 and num2.

+++ lesson_3/task_1/Calculator.h
@@ -5,8 +5,8 @@
 class Calculator
 {
 private:
-    int num1;
-    int num2;
+    double num1;
+    double num2;
 
 public:
     Calculator();
